refactor admin settings sections
